Move constants into own module
