+++ p2panda-rs/src/test_utils/mod.rs
@@ -9,9 +9,8 @@
 //! - mock node and client with experimental materialisation logic implemented
 //! - methods for generating test data (used in `p2panda-js` tests)
 
+pub mod constants;
 pub mod fixtures;
 pub mod mocks;
 pub mod test_data;
-mod utils;
-
-pub use utils::*;
+pub mod utils;
