+++ p2panda-rs/src/test_utils/test_data/json_data.rs
@@ -10,9 +10,9 @@ use crate::hash::Hash;
 use crate::identity::Author;
 use crate::operation::{Operation, OperationEncoded};
 
+use crate::test_utils::constants::DEFAULT_SCHEMA_HASH;
 use crate::test_utils::mocks::Client;
 use crate::test_utils::mocks::Node;
-use crate::test_utils::utils::DEFAULT_SCHEMA_HASH;
 
 /// Next entry args formatted correctly for the test data.
 #[derive(Serialize, Debug)]
